fix: book harness spend on a failed brief and bind the run in one phase
`run_brief` drains and books its cost records in a `finally`, so a brief
that raises still reports what it spent; `_released_tools` binds inside
its own `try` so a failure between bind and release cannot strand a
lease.

`AgentEngine.run()` delegates provider, identity and sampling to
`_bind_run`, which commits every swap together with the identity that
justified it. The provider-resolution helpers move to the run mixin
beside it, leaving the errors mixin to the fatal boundary alone.

The promotion states both readings of the win count: 12 of 15 cells, and
11 of the 13 where either loop cleared the correctness gate.
