+++ data/runtime_stats.yaml
@@ -1,6 +1,6 @@
 schema_version: 1
-last_generated_utc: '2026-08-12T16:11:00Z'
-generator_revision: 272d69474
+last_generated_utc: '2026-08-12T16:41:17Z'
+generator_revision: 5019cf593
 stats:
   tests:
     raw: 46020
